feat(data-subscriber): moved AWS utilities to their own script for reuse
Refs #489
